syncobj: use eventfd instead of stalling fd checks
use eventfd and add it to the event loop and when it recieves a signal
release the buffer, this means we dont stall entire compositor when
waiting for materilisation of the fd. and change its related usage, this
also means we need to store release points that can popup in a container
and signal them all on buffer release.

im not sure why directscanout previously manually tried to signal
releasepoints. remove that and let the buffers releaser handle it.